Make a WinCard that uses Card
Maybe we want to display WinCards is many spots - and don't want to repeat that markup all of the time.

Can make a WinCard that takes in a win as a prop - and have that make use of the Card component.

We then keep a really generic/reusable Card component, and for a more specific use-case - can build off of that.